+++ produtos.php
@@ -106,7 +106,13 @@ if (isset($_GET["a"])) {
 
 		$id = $_POST["id"];
 
-		$res = $db->_exec("DELETE FROM produtos WHERE idProduto = '{$id}'");
+		$sel = $db->select("SELECT idProduto FROM itens_pedido WHERE idProduto = $id");
+
+        if($sel==NULL){
+		    $res = $db->_exec("DELETE FROM produtos WHERE idProduto = '{$id}'");
+        }else{
+            $res=2;
+        }   
 
 		echo $res;
 	}
@@ -285,10 +291,12 @@ include('navbar.php');
                     id: id,
                 },
 		    	success: function retorno_ajax(retorno) {
-                    if(retorno){
+                    if(retorno==1){
 						location.reload();
                     	lista_itens();  
-                	}else{
+                	}else if(retorno==2){
+						alert("Não é possível excluir o produto pois ele está vinculado a um pedido!")
+					}else{	
                     	alert("ERRO AO DELETAR USUÁRIO! " + retorno);
                 	}
 		    	}
